+++ src/simulation/crystalPopulationRepository.h
@@ -9,15 +9,16 @@ class CrystalPopulationRepository
 {
 public:
     CrystalPopulationRepository();
-    void AddDefault();
-    void Remove(unsigned int index);
-    CrystalPopulation &Get(unsigned int index);
-    double GetProbability(unsigned int index) const;
-    unsigned int GetWeight(unsigned int index) const;
-    void SetWeight(unsigned int index, unsigned int weight);
-    unsigned int GetCount() const;
+    void add(CrystalPopulationPreset preset = Random);
+    void remove(unsigned int index);
+    CrystalPopulation &get(unsigned int index);
+    double getProbability(unsigned int index) const;
+    unsigned int getWeight(unsigned int index) const;
+    void setWeight(unsigned int index, unsigned int weight);
+    unsigned int getCount() const;
 
 private:
+    void addDefaults();
     std::vector<CrystalPopulation> mCrystals;
     std::vector<unsigned int> mWeights;
 };