+++ src/simulation/lightSource.h
@@ -7,6 +7,8 @@ struct LightSource
 {
     float altitude;
     float diameter;
+
+    static LightSource createDefaultLightSource();
 };
 
 } // namespace HaloSim